+++ src/main/java/com/teragrep/akv_01/event/EventImpl.java
@@ -45,15 +45,14 @@
  */
 package com.teragrep.akv_01.event;
 
-import com.teragrep.akv_01.time.EnqueuedTime;
-import com.teragrep.akv_01.Parseable;
+import com.teragrep.akv_01.EnqueuedTime;
+import com.teragrep.akv_01.time.EnqueuedTimeImpl;
 import jakarta.json.Json;
 import jakarta.json.JsonReader;
 import jakarta.json.JsonStructure;
 import jakarta.json.stream.JsonParsingException;
 
 import java.io.StringReader;
-import java.time.ZonedDateTime;
 import java.util.Map;
 import java.util.Objects;
 
@@ -63,7 +62,7 @@ public final class EventImpl implements Event {
     private final Map<String, Object> partitionCtx;
     private final Map<String, Object> properties;
     private final Map<String, Object> systemProperties;
-    private final Parseable<ZonedDateTime> enqueuedTimeUtc;
+    private final EnqueuedTime enqueuedTimeUtc;
     private final String offset;
 
     public EventImpl(
@@ -74,7 +73,7 @@ public final class EventImpl implements Event {
             final Object enqueuedTimeUtc,
             final String offset
     ) {
-        this(payload, partitionCtx, properties, systemProperties, new EnqueuedTime(enqueuedTimeUtc), offset);
+        this(payload, partitionCtx, properties, systemProperties, new EnqueuedTimeImpl(enqueuedTimeUtc), offset);
     }
 
     public EventImpl(
@@ -82,7 +81,7 @@ public final class EventImpl implements Event {
             final Map<String, Object> partitionCtx,
             final Map<String, Object> properties,
             final Map<String, Object> systemProperties,
-            final Parseable<ZonedDateTime> enqueuedTimeUtc,
+            final EnqueuedTime enqueuedTimeUtc,
             final String offset
     ) {
         this.payload = payload;
@@ -128,7 +127,7 @@ public final class EventImpl implements Event {
         return systemProperties;
     }
 
-    public Parseable<ZonedDateTime> enqueuedTimeUtc() {
+    public EnqueuedTime enqueuedTimeUtc() {
         return enqueuedTimeUtc;
     }
