+++ src/main/java/com/teragrep/akv_01/event/ParsedEvent.java
@@ -45,10 +45,9 @@
  */
 package com.teragrep.akv_01.event;
 
-import com.teragrep.akv_01.Parseable;
+import com.teragrep.akv_01.EnqueuedTime;
 import jakarta.json.JsonStructure;
 
-import java.time.ZonedDateTime;
 import java.util.Map;
 
 public interface ParsedEvent {
@@ -67,7 +66,7 @@ public interface ParsedEvent {
 
     public abstract Map<String, Object> systemProperties();
 
-    public abstract Parseable<ZonedDateTime> enqueuedTime();
+    public abstract EnqueuedTime enqueuedTime();
 
     public abstract String offset();
 }
