+++ src/main/java/com/teragrep/akv_01/event/PlainEvent.java
@@ -45,10 +45,9 @@
  */
 package com.teragrep.akv_01.event;
 
-import com.teragrep.akv_01.Parseable;
+import com.teragrep.akv_01.EnqueuedTime;
 import jakarta.json.JsonStructure;
 
-import java.time.ZonedDateTime;
 import java.util.Map;
 import java.util.Objects;
 
@@ -98,7 +97,7 @@ public final class PlainEvent implements ParsedEvent {
     }
 
     @Override
-    public Parseable<ZonedDateTime> enqueuedTime() {
+    public EnqueuedTime enqueuedTime() {
         return event.enqueuedTimeUtc();
     }
