+++ src/main/java/com/teragrep/akv_01/time/EnqueuedTimeImpl.java
@@ -45,35 +45,25 @@
  */
 package com.teragrep.akv_01.time;
 
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
+import com.teragrep.akv_01.EnqueuedTime;
 
 import java.time.ZonedDateTime;
-import java.time.format.DateTimeParseException;
 
-public final class EnqueuedTimeTest {
+public final class EnqueuedTimeImpl implements EnqueuedTime {
 
-    @Test
-    void testExpectedCase() {
-        final String origin = "2010-01-01T00:00";
-        final EnqueuedTime enqueuedTime = new EnqueuedTime(origin);
-        Assertions.assertEquals(origin, enqueuedTime.toString());
-        Assertions.assertEquals(ZonedDateTime.parse(origin + "Z"), enqueuedTime.parsed());
+    private final Object origin;
+
+    public EnqueuedTimeImpl(final Object origin) {
+        this.origin = origin;
     }
 
-    @Test
-    void testWithZSuffix() {
-        final String origin = "2010-01-01T00:00Z";
-        final EnqueuedTime enqueuedTime = new EnqueuedTime(origin);
-        Assertions.assertEquals(origin, enqueuedTime.toString());
-        Assertions.assertThrows(DateTimeParseException.class, enqueuedTime::parsed);
+    @Override
+    public ZonedDateTime parsed() {
+        return ZonedDateTime.parse(origin.toString() + "Z");
     }
 
-    @Test
-    void testUnexpectedFormatCase() {
-        final String origin = "01-01-2010T00:00";
-        final EnqueuedTime enqueuedTime = new EnqueuedTime(origin);
-        Assertions.assertEquals(origin, enqueuedTime.toString());
-        Assertions.assertThrows(DateTimeParseException.class, enqueuedTime::parsed);
+    @Override
+    public String toString() {
+        return String.valueOf(origin);
     }
 }
